refact: create lib/charts
